requirements.txt
yfinance>=0.2.40
pandas>=2.0
requests>=2.31
feedparser>=6.0.11
python-dateutil>=2.8
APScheduler>=3.10
streamlit>=1.35
plotly>=5.20
pyyaml>=6.0
